feat(security): siklus hidup data pribadi dan kode dinamis (form keamanan 7.3 dan 4.6)
7.3: penyapu retensi (Penyapu_retensi) yang sebelumnya tidak ada: snapshot
SIMPERUM kedaluwarsa (kecuali yang dirujuk penilaian terkirim), penghitung laju,
token surel, langganan push nonaktif, log lama, audit sangat lama. Berjalan
sekali per 24 jam sesudah respons terkirim (hosting tanpa cron) atau lewat
`php index.php retensi jalankan [kering]`. Hapus akun kini menyapu seluruh
berkas KKN/Magang dan draf penilaian warga (Data_erasure) serta menyamarkan
surel di jejak audit dan forum. Akses staf ke berkas privat, profil warga
terdekripsi, dan NPWP/detail SRP2 dicatat di jejak audit. Register aliran data
keluar dan pemeriksaan skema-vs-kebijakan penghapusan/ekspor sebagai tes.

4.6: inventaris fitur kode dinamis (nol eval/unserialize/include dinamis;
489 pemanggilan variabel semuanya closure lokal). Diperbaiki: XSS lewat
isian ulang di ekspresi Alpine onboarding, 26 echo mentah di skrip inline
(koordinat API pihak ketiga, angka statistik, kode program), model mati
Buka_peta (extract + CRUD tabel dinamis) dihapus. Penjaga tests/dynamic_code_test.php.

Tanpa migrasi. Tes: data_lifecycle_db_test (150), data_lifecycle_test (112),
dynamic_code_test (16), diuji mutasi. Dokumen: docs/engineering/SIKLUS_HIDUP_DATA.md
dan KODE_DINAMIS.md.

+++ application/views/pages/perumahan/detail_perumahan.php
@@ -364,8 +364,8 @@
         $row['lat']  = isset($koordinat[0]) ? trim($koordinat[0]) : '-7.0051';
         $row['long'] = isset($koordinat[1]) ? trim($koordinat[1]) : '110.4381';
         ?>
-        let lat = <?= isset($row['lat']) && !empty($row['lat']) ? $row['lat'] : '-7.0051' ?>; 
-        let lng = <?= isset($row['long']) && !empty($row['long']) ? $row['long'] : '110.4381' ?>; 
+        let lat = <?= (float) (isset($row['lat']) && !empty($row['lat']) ? $row['lat'] : '-7.0051') ?>; 
+        let lng = <?= (float) (isset($row['long']) && !empty($row['long']) ? $row['long'] : '110.4381') ?>; 
         let namaPerumahan = "<?= isset($row['namaPerumahan']) ? htmlspecialchars($row['namaPerumahan']) : 'Lokasi Perumahan' ?>";
 
         // 1. Inisialisasi Kontrol Peta
